fix(perf): resolve review blockers
