+++ crates/zccache-daemon-core/src/daemon/process_tests.rs
@@ -279,17 +279,17 @@ fn auto_priority_decision_normal_on_interactive_when_idle() {
 
 #[test]
 fn in_flight_ticket_returns_pre_increment_count_atomically() {
-    let baseline = current_in_flight_compiles();
+    let baseline = IN_FLIGHT_COMPILES.load(Ordering::Acquire);
     let t1 = InFlightCompileTicket::acquire();
     assert_eq!(t1.in_flight_before(), baseline);
-    assert_eq!(current_in_flight_compiles(), baseline + 1);
+    assert_eq!(IN_FLIGHT_COMPILES.load(Ordering::Acquire), baseline + 1);
     let t2 = InFlightCompileTicket::acquire();
     assert_eq!(t2.in_flight_before(), baseline + 1);
-    assert_eq!(current_in_flight_compiles(), baseline + 2);
+    assert_eq!(IN_FLIGHT_COMPILES.load(Ordering::Acquire), baseline + 2);
     drop(t2);
-    assert_eq!(current_in_flight_compiles(), baseline + 1);
+    assert_eq!(IN_FLIGHT_COMPILES.load(Ordering::Acquire), baseline + 1);
     drop(t1);
-    assert_eq!(current_in_flight_compiles(), baseline);
+    assert_eq!(IN_FLIGHT_COMPILES.load(Ordering::Acquire), baseline);
 }
 
 /// zccache#924: serialize tests that touch the process-wide host
@@ -393,19 +393,6 @@ fn host_counter_saturates_without_overflow() {
     );
     let decision = CompilePriority::Auto.resolve_with_cpu_usage_and_ci(Some(10.0), false, summed);
     assert_eq!(decision.effective, CompilePriority::Low);
-}
-
-#[test]
-fn auto_priority_can_sample_current_load() {
-    let decision = CompilePriority::Auto.resolve_for_current_load();
-    assert_eq!(decision.requested, CompilePriority::Auto);
-    assert!(matches!(
-        decision.effective,
-        CompilePriority::Normal | CompilePriority::Low
-    ));
-    if let Some(cpu_usage_percent) = decision.cpu_usage_percent {
-        assert!((0.0..=100.0).contains(&cpu_usage_percent));
-    }
 }
 
 #[test]
